wsl2にFlutterの設定を追加

+++ setup/init/init_wsl.sh
@@ -6,5 +6,9 @@ echo "# ------------------------------------"
 echo ""
 
 # ルートCAの証明書をinstall
-sudo apt-get update -y
-sudo apt-get install ca-certificates -y
+if [ -f /proc/sys/fs/binfmt_misc/WSLInterop ]; then
+  sudo apt-get update -y
+  sudo apt-get install ca-certificates -y
+else
+  echo "Other os"
+fi
